Adding version 0.3
The main difference here is that  is now pinned to a particular output type. That is, the output type is a type parameter of the  struct itself rather than the  method. This prevents buggy use in the multidimensional case.

Also figured out how to use traits, macros, and const generics to remove hundreds of lines of auto-generated code.